extract shared MapContainer, add tests
