+++ editor.planx.uk/src/@planx/components/FindProperty/Public/Map.tsx
@@ -1,16 +1,21 @@
 import Box from "@mui/material/Box";
 import Link from "@mui/material/Link";
 import Typography from "@mui/material/Typography";
+import styled from "@mui/styles/styled";
 import { visuallyHidden } from "@mui/utils";
 import Card from "@planx/components/shared/Preview/Card";
+import {
+  MapContainer,
+  MapFooter,
+} from "@planx/components/shared/Preview/MapContainer";
 import QuestionHeader from "@planx/components/shared/Preview/QuestionHeader";
+import { useStore } from "pages/FlowEditor/lib/store";
 import React, { useEffect, useState } from "react";
 import { TeamSettings } from "types";
 import Input from "ui/Input";
 import InputLabel from "ui/InputLabel";
 
 import { DEFAULT_NEW_ADDRESS_TITLE, SiteAddress } from "../model";
-import { MapContainer } from ".";
 
 interface PlotNewAddressProps {
   title?: string;
@@ -22,11 +27,21 @@ interface PlotNewAddressProps {
   setPage: React.Dispatch<React.SetStateAction<"os-address" | "new-address">>;
 }
 
+export const DescriptionInput = styled(Box)(({ theme }) => ({
+  display: "flex",
+  paddingTop: theme.spacing(2),
+  paddingBottom: theme.spacing(2),
+}));
+
 export default function PlotNewAddress(props: PlotNewAddressProps) {
-  const [siteDescription, setSiteDescription] = useState<string | undefined>();
+  const [siteDescription, setSiteDescription] = useState<string | undefined>(
+    props.initialProposedAddress?.title ?? undefined
+  );
   const [proposedAddress, setProposedAddress] = useState<
     SiteAddress | undefined
   >(props.initialProposedAddress ?? undefined);
+
+  const environment = useStore((state) => state.previewEnvironment);
 
   useEffect(() => {
     const geojsonChangeHandler = ({ detail: geojson }: any) => {
@@ -66,7 +81,7 @@ export default function PlotNewAddress(props: PlotNewAddressProps) {
         title={props.title || DEFAULT_NEW_ADDRESS_TITLE}
         description={props.description || ""}
       />
-      <MapContainer>
+      <MapContainer environment={environment} interactive>
         <p style={visuallyHidden}>
           An interactive map centred on the local authority district, showing
           the Ordnance Survey basemap. Click to place a point representing your
@@ -75,6 +90,7 @@ export default function PlotNewAddress(props: PlotNewAddressProps) {
         {/* @ts-ignore */}
         <my-map
           id="plot-new-address-map"
+          data-testid="map-web-component"
           zoom={14}
           drawMode
           drawType="Point"
@@ -83,36 +99,38 @@ export default function PlotNewAddress(props: PlotNewAddressProps) {
           showNorthArrow
           osVectorTilesApiKey={process.env.REACT_APP_ORDNANCE_SURVEY_KEY}
         />
-      </MapContainer>
-      <Typography variant="body2">
-        The coordinate location of your address point is:{" "}
-        <strong>
-          {proposedAddress?.x?.toFixed(5) || 0} Easting (X),{" "}
-          {proposedAddress?.y?.toFixed(5) || 0} Northing (Y)
-        </strong>
-      </Typography>
-      <InputLabel label="Describe this site">
-        <Input
-          name="newAddress"
-          bordered
-          onChange={(e) => {
-            setSiteDescription(e.target.value);
-          }}
-          value={siteDescription}
-          id="newAddressInput"
-        />
-      </InputLabel>
-      <Box sx={{ textAlign: "right" }}>
-        <Link
-          component="button"
-          onClick={() => props.setPage("os-address")}
-          disabled={false}
-        >
+        <MapFooter>
           <Typography variant="body2">
-            I want to select an existing address
+            The coordinate location of your address point is:{" "}
+            <strong>
+              {proposedAddress?.x?.toFixed(5) || 0} Easting (X),{" "}
+              {proposedAddress?.y?.toFixed(5) || 0} Northing (Y)
+            </strong>
           </Typography>
-        </Link>
-      </Box>
+          <Link
+            component="button"
+            onClick={() => props.setPage("os-address")}
+            disabled={false}
+          >
+            <Typography variant="body2">
+              I want to select an existing address
+            </Typography>
+          </Link>
+        </MapFooter>
+      </MapContainer>
+      <DescriptionInput data-testid="new-address-input">
+        <InputLabel label="Describe this site">
+          <Input
+            name="newAddress"
+            bordered
+            onChange={(e) => {
+              setSiteDescription(e.target.value);
+            }}
+            value={siteDescription}
+            id="newAddressInput"
+          />
+        </InputLabel>
+      </DescriptionInput>
     </Card>
   );
 }
